Refactor profile screen into a single scroll surface with a floating toolbar.
Separates identity, actions, and trip history so the profile tab is easier to read and maintain without nested scroll regions.
